requirements.txt
pypinyin==0.53.0
pytorch_msssim==1.0.0
blobfile==3.0.0
torch>=2.1.0
timm
mypy
numpy
einops
ftfy
scipy
scikit-image
swanlab
torch==2.1.2
torchvision==0.16.2
torchaudio==2.1.2
transformers==4.43.4
torchvision==0.16.2
transformers==4.43.4 
datasets==2.20.0 
tokenizers 
sentencepiece==0.1.99 
shortuuid
accelerate==0.31.0 
peft 
bitsandbytes
pydantic 
markdown2[all] 
numpy 
scikit-learn==1.2.2
gradio==4.16.0 
gradio_client==0.8.1
requests 
httpx==0.24.0 
uvicorn 
fastapi
einops==0.6.1 
einops-exts==0.0.4
omegaconf==2.1.1
pytorch_lightning==2.2.4 
lightning
tensorboardX 
huggingface-hub==0.24.5 
pillow==10.4.0 
timm==1.0.9
deepspeed==0.12.6 
ninja 
wandb
albumentations==0.4.3